Reorganize files properly
