+++ packages/twenty-front/src/pages/settings/domains/SettingsDomains.tsx
@@ -8,7 +8,7 @@ import { SubMenuTopBarContainer } from '@/ui/layout/page/components/SubMenuTopBa
 import { H2Title } from 'twenty-ui/display';
 import { Section } from 'twenty-ui/layout';
 import { getSettingsPath } from '~/utils/navigation/getSettingsPath';
-import { SettingsWorkspaceDomainCard } from './components';
+import { SettingsWorkspaceDomainCard } from '@/settings/domains/components/SettingsWorkspaceDomainCard';
 
 const StyledMainContent = styled.div`
   display: flex;
