--- packages/twenty-front/src/pages/settings/domains/components/index.ts
@@ -1,1 +1,0 @@
-export { SettingsWorkspaceDomainCard } from './SettingsWorkspaceDomainCard';
