--- packages/twenty-front/src/pages/settings/domains/index.ts
@@ -1,1 +1,0 @@
-export { SettingsDomains } from './SettingsDomains';
